likuor-backend/requirements.txt
argcomplete==3.5.3
asgiref==3.8.1
boto3==1.36.18
botocore==1.36.18
certifi==2025.1.31
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
durationpy==0.9
engineering-notation==0.10.0
hjson==3.1.0
idna==3.10
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
placebo==0.9.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stringify==0.1.1
text-unidecode==1.3
tk==0.1.0
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
zappa==0.59.0
